+++ lib/src/strategies/switch_layout_strategy.dart
@@ -11,6 +11,7 @@ import '../widgets/floating_devices/switch_dev_float.dart';
 import '../widgets/floating_devices/host_dev_float.dart';
 import '../widgets/floating_devices/agent_dev_float.dart';
 import '../widgets/floating_devices/unknown_dev_float.dart';
+import '../widgets/floating_devices/switch_unknown_dev_float.dart';
 import '../widgets/center_device_widget.dart';
 import 'device_layout_strategy.dart';
 
@@ -537,6 +538,18 @@ class SwitchLayoutStrategy extends DeviceLayoutStrategy {
     switch (dev.deviceType) {
       case 'Switch':
         return SwitchDevFloat(
+          portstatus: dev.connectionStatus,
+          position: pd.position,
+          label: label,
+          size: pd.size,
+          connectedPortNum: portNum,
+          deviceStatus: isConfig ? true : dev.deviceStatus,
+          inboundUtilization: inboundUtilization,
+          outboundUtilization: outboundUtilization,
+          isRealDevice: isReal,
+        );
+      case 'SwitchUnknown':
+        return SwitchUnknownDevFloat(
           portstatus: dev.connectionStatus,
           position: pd.position,
           label: label,